Add change name form, routes, controller (#56)
* Add change name form, routes, controller

The current code contains logic for changing the form name along with
form submission email.

This commit builds on the name changing code, with the aim of making it
more production ready.

The following changes have been made:

- splitting the form name change into it's own screen, inline with with
  prototype
- Adds validation, updates page titles on errors and follows the design
  system
- Adds a simple 404 page when a non-existent form is edited
- Moves text strings into locale files for consistency and ease of
  updating
- Adds a simple unit test
- Adds more structure, using the [form object
  pattern](https://jaryl.medium.com/disciplined-rails-form-object-techniques-patterns-part-1-23cfffcaf429)
  and a new Forms:: namespace to keep the size of the model/controllers
  in check
- forms controller broken out a bit into features, with a
  base_controller to cover the common functionality
- Adds 'govuk-components' to make future forms easier to implement

Most of this work is adding more structure to the project to keep
complexity down for future features. This can seem like overkill for
changing the name of a form, but things get hairy quite quickly without
it. Form Objects help keep validation code etc. outside of the model
object and help split controllers up.

There is a bit of an issue with naming - form objects are used to build the
user's form - I'm not sure there is a better name for either concept.
Adding Pages will make things even more fun.

There is still work to do:
- Improve the unit testing - only a single case is covered. Further
  tests would mock the Form model to check things are updated as
  expected
- Add e2e tests, probably using the existing work on cypress
- Update the submission email route to match the name change and clean
  the current forms_controller up a bit
- Improve routing
- Logging

* Add request tests

* Create users before each request spec

* Remove webmock from tests

Co-authored-by: Tom Iles <tom.iles@engineering.digital.dwp.gov.uk>
Co-authored-by: Dan Burnley <dan@danops.uk>