Switch summary list markup to govuk-component
`govuk-component` gem provides us with a summary list component.
Theres a couple of benefits here in that the govuk-component version
should always match the implementation of govuk-frontend nunjucks version.
This means that we dont have to worry about manually updating our version
if govuk-frontend updates their components and those changes ends up
breaking our one.

Apart for the key values being made bold and the actions being on one line,
I think govuk-component implementation does a better job with the actions
because now the screenreader would alert the user to the fact that there
is a list of actions rather than not realising that there is more actions
without navigation through the page.