Update edit-page text to match current content
Add new content for the edit-page.

The design from the protoype uses hint text to describe the radio
buttons. The DFE-govuk component uses the description and makes the
radio items bold. I'll check with the designers if they are happy with
this change or would rather have the hint text as shown in the protoype.

The design system itself doesn't mention anything about using
descriptions, though the markup is essentially the same, it seems to be
only a cosmetic change.