Update edit page text (#99)
Add new content for the edit-page.

The design from the prototype uses hint text to describe the radio
buttons. The DFE-govuk component uses the description and makes the
radio items bold. The designers are happy with
this change.

The translation strings have also been tidied up a bit - there were a few 
unused and duplicate entries.

Co-authored-by: Tom Iles <tom.iles@engineering.digital.dwp.gov.uk>