Make /preview work in dev
It would be nice to be able to share component previews with other
members of the team. Component previews are currently only available
in local development.

This commit adds HostingEnvironment and uses the PAAS_ENVIRONMENT
variable currently set in when deploying the projects.

The current values of PAAS_ENVIRONMENT are:
- dev
- staging
- production

This variable, and RAILS_ENV are used to decide if we are in a test
environment, which is then used to show previews. It's been done this
way because this value is generally useful in other places, allows
testing and doesn't scatter dependencies on PAAS throughout the
codebase.

The ClimateControl gem has been added so that we can test the logic by
changing the ENV.