Add tasklist component (#107)
* Bump govuk_component versions

In order to use the latest view_component version, I've bumped the
versions of the DFE govuk compoents and the govuk styles and js to
match.

* Remove useless view_component line from Gemfile

* Add component setup for rspec, previews

Access a preview of components on /preview

* Add a tasklist view component

Based on the one from the resgister a trainee teacher project, with the
GDS pattern.

* Fix typo for status in locale

Co-authored-by: David Biddle <david.biddle@digital.cabinet-office.gov.uk>

* Normalize locales

* Change connot to cannot in tasklist preview

* Fix gem update

Previous update to Gemfile was too wide ranging. This update only
changes govuk-component and dependency versions.

* Make /preview work in dev

It would be nice to be able to share component previews with other
members of the team. Component previews are currently only available
in local development.

This commit adds HostingEnvironment and uses the PAAS_ENVIRONMENT
variable currently set in when deploying the projects.

The current values of PAAS_ENVIRONMENT are:
- dev
- staging
- production

This variable, and RAILS_ENV are used to decide if we are in a test
environment, which is then used to show previews. It's been done this
way because this value is generally useful in other places, allows
testing and doesn't scatter dependencies on PAAS throughout the
codebase.

The ClimateControl gem has been added so that we can test the logic by
changing the ENV.

Co-authored-by: Tom Iles <tom.iles@engineering.digital.dwp.gov.uk>
Co-authored-by: David Biddle <david.biddle@digital.cabinet-office.gov.uk>