Stop page titles being double html escaped
When editing pages, the title of the page uses question_text. When the
question_text contains some characters, such as ', rails escapes them
twice, which produces titles with hex codes in.

This commit stops that happening and adds a view test to the edit page
view. There are other tests which could be added - testing the title
helper etc. but the view test seemed like the most reliable way to test
the escaping behaviour. This means rendering using the application
layout, which breaks out of the unit testing model a bit.