Remove validation for "What happens next" to make it optional
We have decided that "What happens next" should be optional for the time
being. We will need to get users to manually add in this information to
any existing forms. By removing this validation we also enable them to
"delete/clear" a previous saved content relating to what happens next
if they decide not to have any copy at all.