Add aarch64-linux to supported platforms
Adding `aarch64-linux` to the supported platforms in Gemfile.lock
ensures that `bundle install` installs the correct version of Nokogiri
for the linux container.

Add setting the `GOVUK_APP_DOMAIN` env var if we're pre-compiling the
assets for production because otherwise the operation fails.

Set specific version of bunlder in Dockerfile to match that used in the
Gemfile.lock. We should look at better ways of setting bundler version
in future.