Refactor form task list out to a service
It seems like the time is right to refactor the
private method in the forms controller and better
to make it a service. The idea is that if you pass
in a form, the service will work out which sections
and tasks to return. This array of objects represents
a group of actions that a user can take on a give form.

This refactor will help us when we start looking at
including status for each section and it allows us
to add more methods in future like return a list of
complete/incomplete tasks