add ready_for_live? method to form model
Initially it was thought that we would only check if a form is
ready to be made live once the user attempted to make the form live.
Design thought this approach would confuse users and instead we should
still retrict users from being able to even access the page.

I decided to add this validation check to the form model so that it
could be used to check whether the form was ready to be made live
in the UI as well as a validation check on the "Make live" page just
to ensure that we dont have a scenario of a user getting to make live
page and attempt to make it live when the form is invalid because another
user has updated it to make it invalid.