Reduce duplication when passing forms_id to logs
Each controller which had access to form_id passed it to the logs by
overiding the append_info_to_payload method, calling the parents
versions and adding form_id.

This commit removes these and add's form_id to the
application_controller. If it's null, it's not included in the log
output.