Improve request logging (#183)
* Improve request logging

https://trello.com/c/0COanWVj/274-implement-basic-event-logging-for-forms-admin

Currently the logs sent to splunk are in rails default output format -
spread accross multiple lines and in text format.

This commit does the following:
- adds the lograge gem
- changes the request logging format to json
- outputs a single line of json for each request/response
- adds request_id, user details and form id to request logs where available

I looked at a few approaches to this:
- using rails logging without additional gems
- using lograge
- using semantic_logger
- using https://github.com/alphagov/rack-logstasher

Rails offers support for tagging logs and changing the output format but
doesn't make it easy to change the standard request format without doing
what libraries do anyway.

Semantic logger is a more comprehensive solution used by DfE and offers
more control and options and covers a wider use case than just changing
the reuqest logs.

The alphagov rack-logstasher is aimed at Sinatra/rack applications
which don't have rails request logging already.

Lograge was chosen because it's focussed on changing the standard
request logging and nothing else. We may want to change the format of
other logs e.g. `logger.info "log something specific" to be in json
format (like the runner) and adding some standard keys in those too.

I also looked at passing the request_id to the API and changing the API
logs to json. This means changing at lot of the tests in the project so
will split that into another PR.

To keep this commit self-contained only the request logs have been
altered.

* Reduce duplication when passing forms_id to logs

Each controller which had access to form_id passed it to the logs by
overiding the append_info_to_payload method, calling the parents
versions and adding form_id.

This commit removes these and add's form_id to the
application_controller. If it's null, it's not included in the log
output.

* Add guard clause to check form_id is present

Check form_id is present before including it in logs.

Co-authored-by: Tom Iles <tom.iles@engineering.digital.dwp.gov.uk>