Add email confirmation status
Add tasklist status for the email confirmation loop.

There are two email confirmation tasks which need status:
- sending a code
- confirming a code

Forms which already have a submission email set have both status set to
completed.

The second task status is dependent on the first. If the first task is
complete, the confirmation tasks is active and marked as not started.

When the confirmation task is done, both are marked as complete.

If the user tries to change the email to an address which matches the
current submission email address, both tasks are marked as complete.