Email confirmation tasklist (#204)
* Add colon and newline to the submission_email hint

The tasklist email submission hint text show the current submission
email as one long line.

To make it easier to read a colon and newline now seperate the
explanation text and email address.

* Fix link in submission_email_confirmed view

The "Continue creating a form" link when confirming an email was smaller
than other text on the page.

This was caused by a missing <p> tag wrapping the link. It's been
added.

* Add email confirmation status

Add tasklist status for the email confirmation loop.

There are two email confirmation tasks which need status:
- sending a code
- confirming a code

Forms which already have a submission email set have both status set to
completed.

The second task status is dependent on the first. If the first task is
complete, the confirmation tasks is active and marked as not started.

When the confirmation task is done, both are marked as complete.

If the user tries to change the email to an address which matches the
current submission email address, both tasks are marked as complete.

Co-authored-by: Tom Iles <tom.iles@engineering.digital.dwp.gov.uk>