WIP: spike for adding feature flags to paas envs
Add a way to set feature flags in PAAS deployments, without settings
environment variables during CF deployments.

This could be a good idea or a bad idea, I'm not sure.

It uses the existing PAAS_ENVIRONMENT variable which is already set
when deploying to import an extra settings file, which can add
deployment specific settings. This would make it easy to change feature
flags, mostly for the dev environment as staging and production should
match each other.

There doesn't seem to a nice way of adding extra files in the config
(maybe for good reason?) So it's added after and config reloaded.