Add new show view for live forms
Forms which are live have should be displayed to form creators
differently than forms which have not been made live.

This commit adds a new view for live forms which is only enabled when
the feature flag is true.

When the flag is enabled, live forms will be shown differently - here
ther a placeholder content for this case. The tasklist view can still be
shown to the user by settings the `edit` query param.

The decision to show the live_view is made in the Forms controller and a
new template, live_view is rendered.