Add content for show_live_form
Add the content for showing a live form.

This display is meant to be read only, although it has the option to
edit the form through links. This is to aid the transition to the
draft/live workflow where live forms will be read only and all editing
is done on a draft version of the form, which is then published.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>
Co-authored-by: Alistair Laing <aliuk2012@users.noreply.github.com>