Ignore presence validation if draft/live feature enabled
Theres a lot of opportunity to factor these validations
when the draft/live feature has been completed.
We only added these presence validation initially
to ensure users were not remove the attribute
value once a form went live. The reason is that
there wasn't a "lock" in version of the live form
so they were direcly editting the one and only form record