Move page_list#update to form_controller#mark_pages_section_completed
Initially, we were going to leave this for the time being because potentially
this could also be tidied up further by adding a new endpoint to forms-api
that only concerns about settings sections completed.

After the initial review of the refactor PR, it seemed like it was still
confusing as to why we had `page_list_controller` just for `update` method
and again this is confusing because the update isn't really updating the
pages attribute but an attribute on the forms table. It seems like more
sense to have this action in the forms_controller because it only affects
forms attributes.