refactor mark_completed_form
The controller was doing to much which made testing it completely very
hard to do. Now most of the responsiblity of updating /validating etc
is done in the form object new save method. This makes it easier to test
the form object and let the controller only be concerned whether that
method fails or not and either redirect the user to the next step or
return them to the last page they are on with error message.