Update 'Edit your form' task list page
Updates the content and design of the 'edit your form' page to match the
designs in Mural [1].

When a form has never been published before we what to explain to the
user that they need to make their form live. However, if the form has
been made live for the first task we need to inform the user to make
their changes live.

Although there are some similarities between the copy for the 'create a
form' task list and the 'edit your form' task lists, I found that it was
easier to keep the two separated in the locales file. We make
FormTaskListService responsible for selecting the appropriate content
depending on whether the form has been made live already or not.

In the task list service I wanted to avoid too much repetition, so I
centralised the logic that chooses whether we want content for creating
or editing. The way I (@lfdebrux) did this does result in some fairly
ugly string interpolation; there may be a better way. We also need to
help i18n-tasks match these dynamic keys with what's in the config file,
so we add a new pattern mapper scanner to the config, following the
examples therein [[2]].

Note that these changes don't need to be behind a feature flag, as the
content is valid whether the changes are to a draft or going live
immediately.

This commit also fixes the back link and the location of the preview
link.

[1]: https://app.mural.co/t/gaap0347/m/gaap0347/1667389115580/c92466b787f2ce28bc5ca3786ee60fde382a8aa6?wid=0-1674836853966
[2]: https://github.com/glebm/i18n-tasks/blob/v1.0.11/templates/config/i18n-tasks.yml#L142-L153

Co-authored-by: hannahkc <hannah.cooper@digital.cabinet-office.gov.uk>
Co-authored-by: Alistair Laing <aliuk2012@users.noreply.github.com>
Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>