Refactor Form#status
Now that the form status tag component gets the status text from the
translation file, we no longer need the text in the form model, and we
can simplify the `status` method.

This commit also changes some of the usages of the form status when the
value is enforced by design.