Remove unnecessary copy
This copy is not needed now because when a form is
initially created, it would be a draft. When the user
views the live form, they would see the /live/show page.
When they create/edit draft the form would be live but
now of this content makes sense any more and there is
no need to display a link to their draft form because its
not live.