Update nvmrc to a version nodejs-buildpack supports
GOV.UK PaaS is currently using nodejs-buildpack 1.8.6
https://github.com/cloudfoundry/nodejs-buildpack/releases/tag/v1.8.6

This version doesn't support the lts version 18.15.0 so
We have had to bump it down.

We know that this node version is different from the version in
dockerfile but thats because alphine builds dont have 18.14

Node is only used for asset compiling so we can be a bit more relaxed
about versions.