Refactor how we will store users role.
Instead of using a boolean value to flag if a user
is super admin or not, it might be better to use
enums instead because ruby gives us extra methods

eg.

.super_admin? to check if the user is a super admin
.editor? to check if the user is an editor

.super_admin! to set the user role as super admin
.editor! to set the the user role as editor

User.super_admin to return all users who are super admins
User.editor to return all users who are editors

User.roles will return all the roles as a hash.