Add helpers for testing logins
As we routes which only super_admins should be able to access, we need
to test them in request specs.

This commit adds a few helpers for testing different users.

There are a few questions around how we do it but this is a start.

The current User factory is extended with super_admin and some helpers,
which borrow heavily from here:
https://github.com/alphagov/content-tagger/blob/79047af5d0ac877fc9ad31ea78bc96d823a8eca0/spec/support/authentication_helpers.rb

Co-authored-by: Alistair Laing <aliuk2012@users.noreply.github.com>