Refactor header component
The header component was doing to much than just concentrating on what to display.
It was trying to work out what user profile/signout links to use based on what type
of authentication method the system was using rather than working that out before
calling the component. It meant that the component have dependencies on specific
features in the service to work.

This refactor simplifies the component by abstracting that logic and dependencies
to an application view helper method. Now the component will render what ever is passed
to it. This makes testing easier and should allow us to enhance this component further
in the future.