Update seed.rb to include roles
We only really need to seed for local development. I added this clause so we don't
accidentally end up running this in production.

In addition to updating the default users role to super admin, I also created
a couple extra editors/super-admins so that when the default user visits
`/users` there is at least a couple user accounts to see and to update.