Make UR basic auth user an editor
For the time being, I thought it would be good to
mark the Basic auth as an editor as that role is
mostly what the user will be when they use the system.

We will need to figure out whether we end up
creating additional users in UR env with other
roles.

While testing this locally, I discovered we were
user `current_user` rather than `@current_user`
which would only raise an error if basic auth was
used.