Take control of Warden strategy from gds-sso gem
Allow specifying the Warden strategy in our settings files or with an
environment variable. This replaces relying on the default strategy set
by the gds-sso gem in [[1]].

Note that for the setting, we call it `auth_provider`, not
`auth_strategy`. The term strategy is used consistently in Warden, but
in OmniAuth provider and strategy are used somewhat interchangeably. I
think arguably you could say the provider is the external identity
provider and the strategy is the code you use to communicate with it,
but in the context of a configuration setting there isn't too much
difference in it. To be consistent with what we save in our database
(see PR #469) and to better express the intention of the configuration
variable, we'll go with provider.

[1]: https://github.com/alphagov/gds-sso/blob/a9c2a018dbebd07c35682b1523a2ad27a41d1185/lib/gds-sso.rb#L48-L50