Take control of OmniAuth middleware from gds-sso gem
Override https://github.com/alphagov/gds-sso/blob/a9c2a018dbebd07c35682b1523a2ad27a41d1185/lib/gds-sso.rb#L32-L46.

We need to swap out the middleware in this way so we can add extra
providers to the OmniAuth::Builder later down the line. It's a shame
that we have to copy so much boilerplate from the gem to initialise the
provider, it would be nice if it had more of the defaults hardcoded in
the provider.

Note that we add a `config.after_initialize` callback rather than
calling `config.app_middleware.swap` directly, because we want to make
sure that the swap of the middleware happens after gds-sso has been
initialized.