Add identity provider key to users table
To make switching from one OAuth based identity provider to another a
bit less risky, we add the name of provider to the users table so the
`uid` can be disambiguated.

I'd suggest doing it this way (rather than just getting rid of uid
altogether) because a) it's compatible with what Devise does [[1]] and
b) because then we'll to be able to identify where users in our database
have come from, which seems like it might be useful.

Note that this commit use the name of the provider as OmniAuth knows it,
I thought it would be better this way to allow for simpler code in
future.

[1]: https://github.com/heartcombo/devise/wiki/OmniAuth:-Overview