Add auth provider to User model
To make switching from one OAuth based identity provider to another a
bit less risky, we add the name of provider to the users table so the
`uid` can be disambiguated.

I'd suggest doing it this way (rather than just allowing the uids to be
incosistent, or getting rid of uid altogether) because a) it's
compatible with what Devise does [[1]] and b) because then we'll to be
able to identify where users in our database have come from, which seems
like it might be useful.

This commit also adds a unique index to enforce a constraint; we expect
that uids should be unique for each auth provider, and that is an
important restriction, so let's enforce it.

Note that this commit use the name of the provider as OmniAuth knows it,
gds-sso uses different names for the Warden strategy and OmniAuth
provider, but I think using the OmniAuth provider name allows for
simpler code.

[1]: https://github.com/heartcombo/devise/wiki/OmniAuth:-Overview