Fix 404 returning 500 issue
When a non-existent URL is fetched by a user who is not logged in, the
server throws 500 rather than returning 404 or 403.

This happens because the 404 controller doesn't have a current_user so
throws an exception.

The solution is to skip authentication on all error pages, except the
forbidden page. The forbidden page is only ever shown to authenticated
users - when they don't have permission to see a page. Unauthenticated
users are redirected straight to signon so never get to see this page.