Make the forbidden error message generic
The forbidden error message is shown whenever a pundit policy fails. For
example, it's shown when a user tries to access a from from another
organisation and when en Editor tries to access the /users page.

The current error message was written for the organisation case.

This commit updates the text so that the same page works for both.