Add organisation table
To be able to update a user's organisation within our service we need a
separate organisation table, because `gds-sso` gem will revert any
changes to organisation whenever a user logs in [[1]].

This commit includes a new model, `Organisation`, generated with

    bin/rails generate model Organisation content_id:string:uniq slug:string:uniq name:string

We include the `content_id` in the record because we think we may want
to use the [GOV.UK organisation API][2] going forward (same as Signon
does). Keeping the name (not previously store) is also useful for user
interfaces.

[1]: https://github.com/alphagov/gds-sso/blob/a9c2a018dbebd07c35682b1523a2ad27a41d1185/lib/gds-sso/user.rb#L22
[2]: https://www.api.gov.uk/gds/gov-uk-organisations/#gov-uk-organisations