Show organisation name in user management pages
Show the organisation name (which is nicely formatted), instead of the
organisation slug. Otherwise show "No organisation set".

Note that we don't show the organisation slug anymore, as we are
preparing to switch over to using just the view of organisation within
our service, ignoring whatever GOV.UK Signon says.