Link users to organisations
This commit sets up our models so that users can belong to an
organisation. For users already on our service, we get the organisation
details from the [GOV.UK organisations API][1] and add them to our database
in a migration.

Because Signon includes organisation data in the oauth callback payload,
we might have extra orgs appear in future. We tweak the form policy for
devise so that we can handle users which don't have an associated
organisation, but otherwise don't do anything about this.

In general this commit handles the minimum needed to handle our current
users and organisations. We also don't do anything about organisations
in form objects, as I think some changes to the API will be needed for
that first. Tickets to fill in some of the gaps are already in the
backlog however, so this state of affairs shouldn't persist for too
long.

[1]: https://www.api.gov.uk/gds/gov-uk-organisations/#gov-uk-organisations