Run Tests with bundle exec rspec
For consistency across projects we should run the rspec tests
with the same command so that our CI workflows are simpler to
understand and maintain.

Although running the bin/rake file results in
running the rspec tests via a rake task defined
within the rails-rspec gem it is not clear and took
some time to understand. This change aligns our
workflows on a common and simple approach to running
rspec.