Move creation of test users to authentication feature helpers file
Keep the test behaviour relating to users/permissions in one place, to
help with reasoning about what's happening in tests. Also make it
explicit that we're logging in as a test user before each example,
rather than relying on gds-sso gem to get the first user in the
database.

This also means we no longer need to explicity logout after using
`login_as`.