Revert "Require users to have organisation_id set to use GOV.UK Forms"
This change broke our pipeline, because the migration in
20230426150951_update_organisation_ref_for_existing_users.rb didn't take
for some reason.

We don't need to rollback the migration just yet, because without the
changes in this commit users can still access the service, so for now
we're just rolling back enough to allow the main branch to be deployed
cleanly.