Switch hint input field to textarea
We should look at changing the question hint text input to a textarea.
A few times in research it seems like creators have wanted to put in a
bit more text, or copy paste long bits of hint text that it would be
beneficial for them to see instead of having to scroll a single line input.