Make sure logout spec helper requires user to login again
By default the gds-sso mock strategy will get the first user from the
database, rather than redirecting to the sign in page. If we want to
test whether or not the user is asked to sign in we need to set an extra
environment variable. This commit sets that envvar whenever the logout
helper is called, but resets it to the original setting whenever the
login helper is called, in case the original behaviour is needed for
some reason.