Add code to link OmniAuth and Warden together in our own app
This replaces the gds-sso failure app [[1]] and supersedes its
authentications controller [[2]].

We can have the authentication controller and failure app in one as long
as we a) make the controller useable as an app by adding a `call`
method [[4]] and b) avoid any infinite loops by being careful when we
call Warden.

For redirecting the user back to the page they were trying to access, we
looked at using a scoped session variable, but determined it was
unecessary as we aren't using any Warden scopes. If we do for whatever
reason we can look at the code in Devise for an example [[4]].

For signing out we assume that any provider will have an external sign
out endpoint that needs to be called alongside clearing the client side
information.

[1]: https://github.com/alphagov/gds-sso/blob/a9c2a018dbebd07c35682b1523a2ad27a41d1185/lib/gds-sso/failure_app.rb
[2]: https://github.com/alphagov/gds-sso/blob/a9c2a018dbebd07c35682b1523a2ad27a41d1185/app/controllers/authentications_controller.rb
[3]: https://gist.github.com/dznz/8552522
[4]: https://github.com/heartcombo/devise/blob/4b72064bfcf076478c5c87818b9536b203f6584f/lib/devise/controllers/store_location.rb