Add support for OmniAuth to User model
Bring the code used to look up a user from the OmniAuth hash parameters
from gds-sso [[1]] into our own codebase, so we can reuse it for other
auth providers. We've tweaked it slightly to be more generic, while
still having the interface the gds_sso Warden strategy expects [[2]].

It is expected that `User.find_for_auth` will be generally useful in
future, with Warden strategies passing the desired attributes from the
OmniAuth hash [[3]] or other source of user details to retrieve a user
record.

[1]: https://github.com/alphagov/gds-sso/blob/a9c2a018dbebd07c35682b1523a2ad27a41d1185/lib/gds-sso/user.rb
[2]: https://github.com/alphagov/gds-sso/blob/main/lib/gds-sso/warden_config.rb#L55
[3]: https://github.com/omniauth/omniauth/wiki/Auth-Hash-Schema